Don't overwrite a stash and don't pop no existent stashes.

+++ bin/pop.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 
-echo "Popping stash..."
-
-patch -p0 < ../stash/stash.diff
-rm ../stash/stash.diff
-
-echo "Popped!"
+if [ -f  ../stash/stash.diff ]
+then
+  echo "Popping stash..."
+  patch -p0 < ../stash/stash.diff
+  rm ../stash/stash.diff
+  echo "Popped!"
+else
+  echo "Sorry, no stash file exists"
+fi
